added proper GA files
